Fix tiny format violation
Also updated dependencies.
First commit after almost a year of inactivity. It is kinda strange
how this project is so important to the stianloader toolchain yet
was neglected for such an amount of time. Though I suppose the
project won't degrade in quality if it sits idle for such timespans.

How this discrepancy came into being is a bit beyond me though.